# docs/swagger.yaml
definitions:
  models.AccountAuthResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      refresh_token:
        type: string
    type: object
  models.LoginResponse:
    properties:
      account:
        $ref: '#/definitions/models.AccountAuthResponse'
      err: {}
    type: object
  models.Recipe:
    properties:
      cook:
        type: string
      created_at:
        type: string
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      prep:
        type: string
      updated_at:
        type: string
    type: object
  models.SignupResponse:
    properties:
      account:
        $ref: '#/definitions/models.AccountAuthResponse'
      err: {}
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  schemas.LoginSchemaRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  schemas.RecipeSchemaRequest:
    properties:
      cook:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      prep:
        type: string
    required:
    - name
    type: object
  schemas.SignupSchemaRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: User Login
      parameters:
      - description: user login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.LoginSchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login
      tags:
      - Account
  /logout:
    post:
      consumes:
      - application/json
      description: User Logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SignupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Logout
      tags:
      - Account
  /recipe:
    post:
      consumes:
      - application/json
      description: Create Recipe
      parameters:
      - description: Create recipe payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.RecipeSchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create Recipe
      tags:
      - Recipe
  /recipe/{recipe_id}:
    delete:
      consumes:
      - application/json
      description: Delete Recipe
      parameters:
      - description: Recipe Id
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete Recipe
      tags:
      - Recipe
    get:
      consumes:
      - application/json
      description: Get One Recipe
      parameters:
      - description: Recipe Id
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get One Recipe
      tags:
      - Recipe
    patch:
      consumes:
      - application/json
      description: Update Recipe
      parameters:
      - description: Recipe Id
        in: path
        name: recipe_id
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update Recipe
      tags:
      - Recipe
  /recipes:
    get:
      consumes:
      - application/json
      description: List All Recipes
      parameters:
      - description: pagination offset
        in: query
        name: offset
        type: integer
      - description: pagination size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List All Recipes
      tags:
      - Recipe
  /signup:
    post:
      consumes:
      - application/json
      description: User Signup
      parameters:
      - description: user signup payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.SignupSchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SignupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Signup
      tags:
      - Account
swagger: "2.0"